requirements.txt
python-telegram-bot==20.5
pandas
schedule
